Image-Interpret/requirements.txt
Flask==0.12.2
Werkzeug<0.13.0,>=0.12.0
google-cloud-vision
google-cloud-translate
google-cloud-storage